rename project items to quick access link naming
