+++ Plugins/Flow.Launcher.Plugin.Explorer/Search/QuickAccessLink/QuickAccess.cs
@@ -2,18 +2,18 @@
 using System.Collections.Generic;
 using System.Linq;
 
-namespace Flow.Launcher.Plugin.Explorer.Search.FolderLinks
+namespace Flow.Launcher.Plugin.Explorer.Search.QuickAccessLink
 {
-    public class QuickFolderAccess
+    public class QuickAccess
     {
         private readonly ResultManager resultManager;
 
-        public QuickFolderAccess(PluginInitContext context)
+        public QuickAccess(PluginInitContext context)
         {
             resultManager = new ResultManager(context);
         }
 
-        internal List<Result> FolderListMatched(Query query, List<FolderLink> folderLinks)
+        internal List<Result> FolderListMatched(Query query, List<AccessLink> folderLinks)
         {
             if (string.IsNullOrEmpty(query.Search))
                 return new List<Result>();
@@ -37,7 +37,7 @@ namespace Flow.Launcher.Plugin.Explorer.Search.FolderLinks
                 .ToList();
         }
 
-        internal List<Result> FolderListAll(Query query, List<FolderLink> folderLinks)
+        internal List<Result> FolderListAll(Query query, List<AccessLink> folderLinks)
             => folderLinks
                 .Where(x => x.Type == ResultType.Folder)
                 .Select(item =>
