+++ Plugins/Flow.Launcher.Plugin.Explorer/Settings.cs
@@ -1,5 +1,5 @@
 ﻿using Flow.Launcher.Plugin.Explorer.Search;
-using Flow.Launcher.Plugin.Explorer.Search.FolderLinks;
+using Flow.Launcher.Plugin.Explorer.Search.QuickAccessLink;
 using System.Collections.Generic;
 
 namespace Flow.Launcher.Plugin.Explorer
@@ -8,11 +8,11 @@ namespace Flow.Launcher.Plugin.Explorer
     {
         public int MaxResult { get; set; } = 100;
 
-        public List<FolderLink> QuickFolderAccessLinks { get; set; } = new List<FolderLink>();
+        public List<AccessLink> QuickFolderAccessLinks { get; set; } = new List<AccessLink>();
 
         public bool UseWindowsIndexForDirectorySearch { get; set; } = true;
 
-        public List<FolderLink> IndexSearchExcludedSubdirectoryPaths { get; set; } = new List<FolderLink>();
+        public List<AccessLink> IndexSearchExcludedSubdirectoryPaths { get; set; } = new List<AccessLink>();
 
         public string SearchActionKeyword { get; set; } = Query.GlobalPluginWildcardSign;
